chore: address feedback
- add copyright headers
- abstract AmplifyError types
