+++ packages/auth/src/common/AuthErrorStrings.ts
@@ -1,7 +1,10 @@
-import { AuthErrorMap } from '../errors/types/models';
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { AmplifyErrorMap } from '@aws-amplify/core/src/types/types';
 import { AuthValidationErrorCode } from '../errors/types/validation';
 
-export const validationErrorMap: AuthErrorMap<AuthValidationErrorCode> = {
+export const validationErrorMap: AmplifyErrorMap<AuthValidationErrorCode> = {
 	[AuthValidationErrorCode.EmptyChallengeResponse]: {
 		message: 'challengeResponse is required to confirmSignIn',
 	},
