+++ packages/auth/src/errors/types/validation.ts
@@ -1,3 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 export enum AuthValidationErrorCode {
 	EmptySignInUsername = 'EmptySignInUsername',
 	EmptySignInPassword = 'EmptySignInPassword',
