+++ packages/auth/src/errors/utils/assertServiceError.ts
@@ -1,12 +1,15 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { ServiceError } from '@aws-amplify/core/src/types/types';
 import { AuthError } from '../AuthError';
-import { ServiceError } from '../types/models';
 
 export function assertServiceError(
 	error: unknown
 ): asserts error is ServiceError {
 	if (
 		!error ||
-		!((error as ServiceError).name || !(error as ServiceError).message)
+		!((error as ServiceError).name && (error as ServiceError).message)
 	) {
 		throw new AuthError({
 			name: 'UnknownError',
