+++ packages/auth/src/providers/cognito/types/errors/service.ts
@@ -1,3 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 export enum AssociateSoftwareTokenException {
 	ConcurrentModificationException = 'ConcurrentModificationException',
 	ForbiddenException = 'ForbiddenException',
