+++ packages/core/src/Errors.ts
@@ -14,6 +14,7 @@ export class AmplifyError extends Error {
 	underlyingError?: Error | unknown;
 	recoverySuggestion?: string;
 	/**
+	 *  Constructs an AmplifyError.
 	 *
 	 * @param message text that describes the main problem.
 	 * @param underlyingError the underlying cause of the error.
