+++ packages/core/src/types/types.ts
@@ -1,3 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 import { InputLogEvent, LogGroup } from '@aws-sdk/client-cloudwatch-logs';
 import { Credentials } from '@aws-sdk/types';
 
@@ -69,4 +72,16 @@ export type ErrorParams = {
 	name: string;
 	recoverySuggestion?: string;
 	underlyingError?: Error | unknown;
+};
+
+export type AmplifyErrorMap<ErrorCode extends string> = {
+	[name in ErrorCode]: {
+		message: string;
+		recoverySuggestion?: string;
+	};
+};
+
+export type ServiceError = {
+	name: string;
+	message: string;
 };
